+++ CHANGELOG.md
@@ -6,6 +6,7 @@ NEW FEATURES:
 
 ENHANCEMENTS:
 * Added `tofu test -json` types to website Machine-Readable UI documentation ([1408](https://github.com/opentofu/opentofu/issues/1408))
+* Made `tofu plan` with `generate-config-out` flag replace JSON strings with `jsonencode` functions calls. ([#1595](https://github.com/opentofu/opentofu/pull/1595))
 * Added `pedantic` flag which enables warnings to be treated as errors ([757](https://github.com/opentofu/opentofu/issues/757))
 
 BUG FIXES: